+++ Model/Admin.java
@@ -1,4 +1,4 @@
-package model;
+package Model;
 
 import java.text.SimpleDateFormat;
 import java.util.*;
