+++ Model/IUser.java
@@ -1,4 +1,4 @@
-package model;
+package Model;
 
 public interface IUser {
