+++ Model/Trip.java
@@ -1,4 +1,4 @@
-package model;
+package Model;
 
 import java.util.Date;
 import java.util.ArrayList;
