+++ Model/User.java
@@ -1,4 +1,4 @@
-package model;
+package Model;
 
 import java.util.List;
 import java.util.Date;
